+++ redwood-compose/src/commonTest/kotlin/app/cash/redwood/compose/BackHandlerTest.kt
@@ -34,8 +34,8 @@ import assertk.assertions.isEqualTo
 import assertk.assertions.isInstanceOf
 import assertk.assertions.single
 import com.example.redwood.testing.compose.Text
-import com.example.redwood.testing.widget.TestSchemaTester
-import com.example.redwood.testing.widget.TextValue
+import com.example.redwood.testing.testing.TestSchemaTester
+import com.example.redwood.testing.testing.TextValue
 import kotlin.test.Test
 import kotlinx.coroutines.TimeoutCancellationException
 import kotlinx.coroutines.test.runTest
