+++ redwood-compose/src/commonTest/kotlin/app/cash/redwood/compose/ListeningButton.kt
@@ -18,8 +18,8 @@ package app.cash.redwood.compose
 import app.cash.redwood.Modifier
 import app.cash.redwood.testing.WidgetValue
 import app.cash.redwood.widget.ChangeListener
+import com.example.redwood.testing.testing.ButtonValue
 import com.example.redwood.testing.widget.Button
-import com.example.redwood.testing.widget.ButtonValue
 
 class ListeningButton : Button<WidgetValue>, ChangeListener {
   private val changes = ArrayList<String>()
