+++ redwood-lazylayout-compose/src/commonTest/kotlin/app/cash/redwood/lazylayout/compose/LazyListTest.kt
@@ -16,13 +16,13 @@
 package app.cash.redwood.lazylayout.compose
 
 import app.cash.redwood.Modifier
-import app.cash.redwood.lazylayout.widget.LazyListValue
+import app.cash.redwood.lazylayout.testing.LazyListValue
 import assertk.assertThat
 import assertk.assertions.isEmpty
 import assertk.assertions.isEqualTo
 import com.example.redwood.testing.compose.Text
-import com.example.redwood.testing.widget.TestSchemaTester
-import com.example.redwood.testing.widget.TextValue
+import com.example.redwood.testing.testing.TestSchemaTester
+import com.example.redwood.testing.testing.TextValue
 import kotlin.test.Test
 import kotlinx.coroutines.test.runTest
