+++ test-app/presenter-treehouse/src/jsMain/kotlin/com/example/redwood/testing/treehouse/RealTestAppPresenter.kt
@@ -18,7 +18,7 @@ package com.example.redwood.testing.treehouse
 import app.cash.redwood.treehouse.StandardAppLifecycle
 import app.cash.redwood.treehouse.ZiplineTreehouseUi
 import app.cash.redwood.treehouse.asZiplineTreehouseUi
-import com.example.redwood.testing.compose.TestSchemaProtocolBridge
+import com.example.redwood.testing.protocol.guest.TestSchemaProtocolBridge
 import kotlinx.serialization.json.Json
 
 class RealTestAppPresenter(
